requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic[email]==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.26.0
yfinance==0.2.28
pandas==2.1.4
python-jose[cryptography]
PyJWT